+++ Containers/TMap.h
@@ -15,7 +15,7 @@ typedef struct TMap
 } TMap;
 
 // types of all [...] arguments should be a TRtti. 
-TMap* TMap_Init(TIterator* it, size_t type_Count, size_t value_Count, ...);
+TMap* TMap_Init(TIterator* it, uint32_t type_Count, uint32_t value_Count, ...);
 
 // Returns true if [key] does not exist inside [map->Pairs].
 // Does not add a key-value pair if the provided key already exists inside of the map.
@@ -25,7 +25,7 @@ bool TMap_Add(TMap* map, TPair* key_Value_Pair);
 // Returns true if the added [key]s does not exist inside [map->Pairs].
 // Does not add key-value pairs if the provided keys already exists inside of the map.
 // Copies an added value if it is not a pointer type.
-bool TMap_Multi(TMap* map, size_t value_Count, ...);
+bool TMap_Multi(TMap* map, uint32_t value_Count, ...);
 
 void TMap_Remove(TMap* map, TString* key);
